The cards expired date validation control bux fixed.
The expiry date on the cards should be based on the last day of that month. Since the current structure is based only on the month, it does not work correctly. As a bug fix, the last day of the month was found and added to the card expiry date.