Replaced textcontroller listeners with onchange events
Resolve the issue of validating using old/previous values.

Issue is highlighted when `onCreditCardModelChange` is invoked and when validating separate field using their validator functions.

e.g. 

```dart
void onCreditCardModelChange(CreditCardModel? creditCardModel) {
    bool _formValid = true;
    if (cardNumber.isNotEmpty) {
      if (!(cardNumberKey.currentState?.isValid ?? false)) {
        _formValid = false;
      }
    } else {
      _formValid = false;
    }
  }
```

Upon calling `cardNumberKey.currentState?.isValid`, the `cardNumberTextController` has the new value however since a `Form` is used, the `FormStateField`'s `value` hasn't been updated (via the `Form.save()` function). This means, even though the textController has the new value, the `FormFieldState` key hasn't been updated with the new value resulting in validation (via the `FormStateField`) using the old value. 

This is visible on the UI as the field being "one-step behind" on each text change.

This is consistent with what is expected of Flutter's framework, the problem is with this library.